# docs/openapi.yaml
openapi: '3.0.3'
info:
  title: Centre Commercial API
  description: |
    API de gestion de centre commercial (MEAN stack).

    ## Rôles
    - **ADMIN** : gestion complète (boutiques, catégories, utilisateurs, stats)
    - **SHOP_MANAGER** : gestion de sa propre boutique (produits, promotions)
    - **CUSTOMER** : navigation, favoris, avis

    ## Authentification
    Stratégie double token JWT :
    - **Access token** : 15 min — à envoyer dans `Authorization: Bearer <token>`
    - **Refresh token** : 7 jours — à envoyer dans le body pour rafraîchir

    Les tokens ne sont délivrés qu'après vérification de l'email.
  version: '1.0.0'
servers:
  - url: http://localhost:3001/api
    description: Serveur de développement
security:
  - BearerAuth: []
...
paths:
...
  /myshop/products:
    get:
      tags: [My Products]
      operationId: getMyProducts
      summary: Lister ses produits
      responses:
        '200':
          description: Liste des produits
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
    post:
      tags: [My Products]
      operationId: createProduct
      summary: Créer un produit (boutique doit être ACTIVE)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Produit créé
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '403':
          $ref: '#/components/responses/Forbidden'
  /myshop/products/{id}:
    put:
      tags: [My Products]
      operationId: updateProduct
      summary: Modifier un produit
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Produit mis à jour
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [My Products]
      operationId: deleteProduct
      summary: Supprimer un produit
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Produit supprimé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # MY PROMOTIONS (SHOP_MANAGER)
  # ---------------------------------------------------------------------------
  /myshop/promotions:
    get:
      tags: [My Promotions]
      operationId: getMyPromotions
      summary: Lister ses promotions
      responses:
        '200':
          description: Liste des promotions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Promotion'
    post:
      tags: [My Promotions]
      operationId: createPromotion
      summary: Créer une promotion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionInput'
      responses:
        '201':
          description: Promotion créée
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Promotion'
        '403':
          $ref: '#/components/responses/Forbidden'
  /myshop/promotions/{id}:
    put:
      tags: [My Promotions]
      operationId: updatePromotion
      summary: Modifier une promotion
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionInput'
      responses:
        '200':
          description: Promotion mise à jour
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Promotion'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [My Promotions]
      operationId: deletePromotion
      summary: Supprimer une promotion
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Promotion supprimée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # PUBLIC (pas d'authentification)
  # ---------------------------------------------------------------------------
...
  /customer/favorites/{shopId}:
    post:
      tags: [Customer]
      operationId: addFavorite
      summary: Ajouter une boutique aux favoris
      parameters:
        - { name: shopId, in: path, required: true, schema: { type: string } }
      responses:
        '201':
          description: Ajouté aux favoris
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Favorite'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Customer]
      operationId: removeFavorite
      summary: Retirer une boutique des favoris
      parameters:
        - { name: shopId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Retiré des favoris
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '404':
          $ref: '#/components/responses/NotFound'
  /customer/shops/{shopId}/reviews:
    post:
      tags: [Customer]
      operationId: createReview
      summary: Laisser un avis sur une boutique (upsert — un seul avis par boutique)
      parameters:
        - { name: shopId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewInput'
      responses:
        '201':
          description: Avis enregistré
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # ARTICLES (legacy)
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: ID MongoDB de la ressource
...
  responses:
...
    Forbidden:
      description: Accès interdit — rôle insuffisant ou compte suspendu
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Conflit — ressource déjà existante
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
...
    UserRole:
      type: string
      enum: [CUSTOMER, SHOP_MANAGER, ADMIN]
    UserStatus:
      type: string
      enum: [VALID, SUSPENDED, WAITING]
    ShopStatus:
      type: string
      enum: [PENDING, ACTIVE, SUSPENDED]

    # --- Auth ---
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        address:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
...
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time

    # --- Shops ---
...
    Shop:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        manager:
          $ref: '#/components/schemas/UserProfile'
        category:
          $ref: '#/components/schemas/Category'
        note:
          type: number
          minimum: 0
          maximum: 5
          description: Note moyenne calculée à partir des avis
        hours:
          type: string
        image:
          type: string
        status:
          $ref: '#/components/schemas/ShopStatus'
        createdAt:
          type: string
          format: date-time
...
    ProductInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          minimum: 0
        image:
          type: string
        category:
          type: string
          description: ID de la catégorie
        isActive:
          type: boolean
          default: true
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        image:
          type: string
        shop:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Shop'
        category:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Category'
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # --- Promotions ---
    PromotionInput:
      type: object
      required: [product, discountPercent, startDate, endDate]
      properties:
        product:
          type: string
          description: ID du produit
        discountPercent:
          type: integer
          minimum: 1
          maximum: 100
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        isActive:
          type: boolean
          default: true
    Promotion:
      type: object
      properties:
        id:
          type: string
        product:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Product'
        shop:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Shop'
        discountPercent:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # --- Reviews ---
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          maxLength: 1000
    Review:
      type: object
      properties:
        id:
          type: string
        user:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/UserProfile'
        shop:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        createdAt:
          type: string
          format: date-time

    # --- Favorites ---
    Favorite:
      type: object
      properties:
        id:
          type: string
        user:
          type: string
        shop:
          $ref: '#/components/schemas/Shop'
        createdAt:
          type: string
          format: date-time

    # --- Articles (legacy) ---
...